Update and rename get_imdbtop.py to imdb_top.py
1. Renamed the filename to imdb_top.py to match the name of the function.
2. Input is converted to int inside validate function, so that other validations can also be performed.
3. Added docstring in the function showing that imdb_top_n is an int and the function returns None.
4. Raising ValueError, TypeError etc. based on input.

** I could not figure out doctest for function where user input is expected.
